+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,12 @@
 # playground
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [5aa6631]
+  - @westpac/style-config@1.0.2
+
 ## 0.1.3
 
 ### Patch Changes